[embedlite-components] Improve network logging
When the EMBED_CONSOLE="network" environment variable is set the browser
will log network activity. These changes make it easier to tie the data
downloaded with the request and response headers, including the URL the
data was downloaded from.

In particular, the download URL is now output directly before the data
output.

The change also makes the script work harder to interpret whether the
data is text or not, but relying not just on the encoding but also the
MIME type.